Move InstallFromImageTask to payload.base
This installation task is shared between live_image and live_os. So move
it between shared tasks.
