+++ pyanaconda/modules/payload/base/installation.py
@@ -17,6 +17,7 @@
 #
 from pyanaconda.modules.common.task import Task
 from pyanaconda.modules.common.errors.payload import InstallError
+from pyanaconda.core.constants import INSTALL_TREE
 from pyanaconda.core.util import execWithRedirect
 from pyanaconda.modules.payload.live.utils import create_rescue_image
 
@@ -24,28 +25,44 @@ from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
 
 
-class InstallFromTarTask(Task):
-    """Task to install the payload from tarball."""
+class InstallFromImageTask(Task):
+    """Task to install the payload from image."""
 
-    def __init__(self, tarfile_path, dest_path, kernel_version_list):
+    def __init__(self, dest_path, kernel_version_list, source=None):
+        """Create a new task.
+
+        :param dest_path: installation destination root path
+        :type dest_path: str
+        :param kernel_version_list: list of kernel versions for rescue initrd images
+                                    to be created
+        :type krenel_version_list: list(str)
+        """
         super().__init__()
-        self._tarfile_path = tarfile_path
+        self._source = source
         self._dest_path = dest_path
         self._kernel_version_list = kernel_version_list
 
     @property
     def name(self):
-        return "Install the payload from a tarball"
+        return "Install the payload from image"
 
     def run(self):
-        """Run installation of the payload from a tarball."""
-        cmd = "tar"
-        # preserve: ACL's, xattrs, and SELinux context
-        args = ["--numeric-owner", "--selinux", "--acls", "--xattrs", "--xattrs-include", "*",
-                "--exclude", "dev/*", "--exclude", "proc/*", "--exclude", "tmp/*",
-                "--exclude", "sys/*", "--exclude", "run/*", "--exclude", "boot/*rescue*",
-                "--exclude", "boot/loader", "--exclude", "boot/efi/loader",
-                "--exclude", "etc/machine-id", "-xaf", self._tarfile_path, "-C", self._dest_path]
+        """Run installation of the payload from image."""
+        # TODO: remove this check for None when Live Image payload will support sources
+        # The None check is just a temporary hack that Live OS has source but Live Image don't
+        if self._source is not None and not self._source.is_ready():
+            raise InstallError("Source is not set up!")
+
+        cmd = "rsync"
+        # preserve: permissions, owners, groups, ACL's, xattrs, times,
+        #           symlinks, hardlinks
+        # go recursively, include devices and special files, don't cross
+        # file system boundaries
+        # TODO: source will provide us source path instead of using constant here
+        args = ["-pogAXtlHrDx", "--exclude", "/dev/", "--exclude", "/proc/", "--exclude", "/tmp/*",
+                "--exclude", "/sys/", "--exclude", "/run/", "--exclude", "/boot/*rescue*",
+                "--exclude", "/boot/loader/", "--exclude", "/boot/efi/loader/",
+                "--exclude", "/etc/machine-id", INSTALL_TREE + "/", self._dest_path]
         try:
             rc = execWithRedirect(cmd, args)
         except (OSError, RuntimeError) as e:
@@ -57,7 +74,7 @@ class InstallFromTarTask(Task):
             msg = "%s exited with code %d" % (cmd, rc)
             log.info(msg)
 
-        if err:
+        if err or rc == 11:
             raise InstallError(err or msg)
 
         create_rescue_image(self._dest_path, self._kernel_version_list)
