+++ pyanaconda/modules/payload/payloads/live_image/live_image.py
@@ -31,6 +31,7 @@ from pyanaconda.modules.common.errors.payload import SourceSetupError
 from pyanaconda.modules.payload.base.handler_base import PayloadHandlerBase
 from pyanaconda.modules.payload.base.initialization import CopyDriverDisksFilesTask, \
     UpdateBLSConfigurationTask
+from pyanaconda.modules.payload.base.installation import InstallFromImageTask
 from pyanaconda.modules.payload.payloads.live_image.live_image_interface import \
     LiveImageHandlerInterface
 from pyanaconda.modules.payload.payloads.live_image.initialization import \
@@ -38,7 +39,7 @@ from pyanaconda.modules.payload.payloads.live_image.initialization import \
     TeardownInstallationSourceImageTask
 from pyanaconda.modules.payload.live.utils import get_kernel_version_list, \
     get_kernel_version_list_from_tar, url_target_is_tarfile
-from pyanaconda.modules.payload.live.installation import InstallFromImageTask, InstallFromTarTask
+from pyanaconda.modules.payload.live.installation import InstallFromTarTask
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
